Refactored code, began local testing.
